+++ test.cpp
@@ -5,21 +5,20 @@
 #include <iostream>
 
 int main (int argc, char** argv) {
-	ng::Time time;
+	ng::Event event;
 	ng::Graphics graphics;
 	ng::Audio audio;
-	ng::Event event;
+	ng::Time time;
 	
 	try {
 		ng::init();
-		
-		time.init();
+		event.init(&graphics);
 		graphics.init("Fire Days Demo", 640, 480);
 		audio.init();
-		event.init(&graphics);
-		
+		time.init();
 	} catch (const std::exception& ex) {
-		std::cout << NG_HERE << ": can't init: " << ex.what() << "\n";
+		std::cout << "error at startup:\n"
+			<< ex.what() << "\n";
 		exit(EXIT_FAILURE);
 	}
 	
@@ -27,11 +26,40 @@ int main (int argc, char** argv) {
 	background_color.init(0, 0, 0);
 	ng::Color draw_color;
 	draw_color.init(255, 255, 255);
+	ng::Color color_key;
+	color_key.init(0, 0, 0);
 	ng::Rect rect;
-	rect.init(250, 200, 150, 100);
+	rect.init((graphics.rect.w/2)-256, (graphics.rect.h/2)-80, 256*2, 80*2);
+	
+	ng::Image image;
+	ng::Clip clip;
+	
+	const char* file;
+	try {
+		image.init(&graphics, file="game-data/text.bmp", &color_key);
+		clip.init(&audio, file="game-data/Corncob.wav");
+	} catch (const std::exception& ex) {
+		std::cout << "error loading \"" << file << "\":\n"
+			<< ex.what() << "\n";
+		exit(EXIT_FAILURE);
+	}
 	
 	ng::Channel channel;
 	channel.init();
+	
+	try {
+		channel.play_sound(&clip, ng::SoundLoop);
+	} catch (const std::exception& ex) {
+		std::cout << "error configuring data:\n"
+			<< ex.what() << "\n";
+		exit(EXIT_FAILURE);
+	}
+	
+	int up, down, left, right;
+	up = 0;
+	down = 0;
+	left = 0;
+	right = 0;
 	
 	while (true) {
 		while (event.next()) {
@@ -39,20 +67,77 @@ int main (int argc, char** argv) {
 				goto quit;
 			}
 			
-			// handle events
+			switch (event.mode) {
+				case ng::Quit:
+					goto quit;
+					break;
+				case ng::WindowEvent:
+					rect.x = (graphics.rect.w/2)-(rect.w/2);
+					rect.y = (graphics.rect.h/2)-(rect.h/2);
+					break;
+				case ng::KeyPress: {
+					switch (event.key.scancode) {
+						case SDL_SCANCODE_SPACE:
+							rect.x = (graphics.rect.w/2)-(rect.w/2);
+							rect.y = (graphics.rect.h/2)-(rect.h/2);
+							up = 0;
+							down = 0;
+							left = 0;
+							right = 0;
+							break;
+						case SDL_SCANCODE_W:
+							up = 1;
+							break;
+						case SDL_SCANCODE_S:
+							down = 1;
+							break;
+						case SDL_SCANCODE_A:
+							left = 1;
+							break;
+						case SDL_SCANCODE_D:
+							right = 1;
+							break;
+						default: {}
+					}
+					break;
+				}
+				case ng::KeyRelease: {
+					switch (event.key.scancode) {
+						case SDL_SCANCODE_W:
+							up = 0;
+							break;
+						case SDL_SCANCODE_S:
+							down = 0;
+							break;
+						case SDL_SCANCODE_A:
+							left = 0;
+							break;
+						case SDL_SCANCODE_D:
+							right = 0;
+							break;
+						default: {}
+					}
+					break;
+				}
+				default: {}
+			}
 		}
+		
+		rect.x += right - left;
+		rect.y += down - up;
 		
 		try {
 			graphics.set_color(&background_color);
 			graphics.clear();
 			
 			graphics.set_color(&draw_color);
-			graphics.draw_rect(&rect, ng::DrawFill);
+			graphics.draw_image(&image, &image.rect, &rect);
 			
 			graphics.draw();
 			
 		} catch (const std::exception& ex) {
-			std::cout << NG_HERE << ": graphics error: " << ex.what() << "\n";
+			std::cout << "error drawing graphics:\n"
+				<< ex.what() << "\n";
 			exit(EXIT_FAILURE);
 		}
 		
@@ -62,7 +147,8 @@ int main (int argc, char** argv) {
 			audio.play();
 			
 		} catch (const std::exception& ex) {
-			std::cout << NG_HERE << ": audio error: " << ex.what() << "\n";
+			std::cout << "error playing audio:\n"
+				<< ex.what() << "\n";
 			exit(EXIT_FAILURE);
 		}
 		
@@ -70,9 +156,13 @@ int main (int argc, char** argv) {
 	}
 
 quit:
+	image.quit();
 	graphics.quit();
+	
+	clip.quit();
 	channel.quit();
 	audio.quit();
+	
 	ng::quit();
 	
 	return EXIT_SUCCESS;